riscv: cheri: Retain provenance in tagging/untagging macros
Make sure that the dummy versions of the macros that tag/untag
addresses retain the capability tag. The feature itself is not
compatible with CHERI.

+++ arch/riscv/include/asm/uaccess.h
@@ -29,12 +29,12 @@ static inline unsigned long __untagged_addr_remote(struct mm_struct *mm, unsigne
 }
 
 #define untagged_addr(addr) ({							\
-	unsigned long __addr = (__force unsigned long)(addr);			\
+	uintptr_t __addr = (__force uintptr_t)(addr);			\
 	(__force __typeof__(addr))__untagged_addr_remote(current->mm, __addr);	\
 })
 
 #define untagged_addr_remote(mm, addr) ({					\
-	unsigned long __addr = (__force unsigned long)(addr);			\
+	uintptr_t __addr = (__force uintptr_t)(addr);			\
 	mmap_assert_locked(mm);							\
 	(__force __typeof__(addr))__untagged_addr_remote(mm, __addr);		\
 })
